Created Dropdown helper class for reusability
